refactor: 로또 금액 관련 에러 메세지 수정
