+++ src/test/java/lotto/domain/LottosTest.java
@@ -1,0 +1,10 @@
+package lotto.domain;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+public class LottosTest {
+
+    @Test
+    @DisplayName("로또 발행 금")
+}
